Component with useLoader can use dynamic import
