+++ examples/with-three-js/pages/birds.js
@@ -1,42 +1,6 @@
-import { useRef, useState, useEffect, Suspense } from 'react'
-import * as THREE from 'three'
-import { Canvas, useFrame, useLoader } from 'react-three-fiber'
-
-let GLTFLoader
-
-const Bird = ({ speed, factor, url, ...props }) => {
-  const gltf = useLoader(GLTFLoader, url)
-  const group = useRef()
-  const [mixer] = useState(() => new THREE.AnimationMixer())
-  useEffect(
-    () => void mixer.clipAction(gltf.animations[0], group.current).play(),
-    [gltf.animations, mixer]
-  )
-  useFrame((state, delta) => {
-    group.current.rotation.y +=
-      Math.sin((delta * factor) / 2) * Math.cos((delta * factor) / 2) * 1.5
-    mixer.update(delta * speed)
-  })
-  return (
-    <group ref={group}>
-      <scene name="Scene" {...props}>
-        <mesh
-          name="Object_0"
-          morphTargetDictionary={gltf.__$[1].morphTargetDictionary}
-          morphTargetInfluences={gltf.__$[1].morphTargetInfluences}
-          rotation={[1.5707964611537577, 0, 0]}
-        >
-          <bufferGeometry attach="geometry" {...gltf.__$[1].geometry} />
-          <meshStandardMaterial
-            attach="material"
-            {...gltf.__$[1].material}
-            name="Material_0_COLOR_0"
-          />
-        </mesh>
-      </scene>
-    </group>
-  )
-}
+import { Suspense } from 'react'
+import { Canvas } from 'react-three-fiber'
+const Bird = dynamic(() => import('../components/Bird'), { ssr: false })
 
 const Birds = () => {
   return new Array(5).fill().map((_, i) => {
@@ -65,9 +29,6 @@ const Birds = () => {
 }
 
 const BirdsPage = (props) => {
-  useEffect(() => {
-    GLTFLoader = require('three/examples/jsm/loaders/GLTFLoader').GLTFLoader
-  }, [])
   return (
     <>
       <Canvas camera={{ position: [0, 0, 35] }}>
